requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
alembic==1.13.2
altair==5.4.1
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
attrs==24.2.0
backoff==2.2.1
blinker==1.8.2
boto3==1.35.8
botocore==1.35.8
cachetools==5.5.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
cohere==5.8.1
colorama==0.4.6
colorlog==6.8.2
dataclasses-json==0.6.7
datasets==2.21.0
Deprecated==1.2.14
dill==0.3.8
dirtyjson==1.0.8
distro==1.9.0
dspy-ai==2.4.13
exceptiongroup==1.2.2
fastavro==1.9.5
filelock==3.15.4
frozenlist==1.4.1
fsspec==2024.6.1
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.24.6
idna==3.8
Jinja2==3.1.4
jiter==0.5.0
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
langchain==0.2.16
langchain-community==0.2.16
langchain-core==0.2.38
langchain-text-splitters==0.2.4
langsmith==0.1.116
llama-index-core==0.11.2
llama-parse==0.5.1
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
mdurl==0.1.2
multidict==6.0.5
multiprocess==0.70.16
mypy-extensions==1.0.0
narwhals==1.5.5
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.9.1
numpy==1.26.4
openai==1.42.0
optuna==3.6.1
orjson==3.10.7
packaging==24.1
pandas==2.2.2
parameterized==0.9.0
pillow==10.4.0
protobuf==5.28.0
pyarrow==17.0.0
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
ratelimit==2.2.1
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
rich==13.8.0
rpds-py==0.20.0
s3transfer==0.10.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.32
streamlit==1.38.0
structlog==24.4.0
tenacity==8.5.0
tiktoken==0.7.0
tokenizers==0.20.0
toml==0.10.2
tornado==6.4.1
tqdm==4.66.5
types-requests==2.31.0.6
types-urllib3==1.26.25.14
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
urllib3==1.26.19
watchdog==4.0.2
wrapt==1.16.0
xxhash==3.5.0
yarl==1.9.4
